Corrected small bugs on S-LAM flattening and edition. Made edition more robust. Added a test project. Still a little problem on several-level flattening to solve.
git-svn-id: http://svn.code.sf.net/p/dftools/svn/trunk@99 241f8b49-2e81-4b1e-ac05-70ce8349dc88

+++ eclipse/plugins/net.sf.dftools.architecture/src/net/sf/dftools/architecture/slam/component/ComNode.java
@@ -63,12 +63,12 @@ public interface ComNode extends Component {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Speed</em>' attribute.
-	 * @see #setSpeed(int)
+	 * @see #setSpeed(float)
 	 * @see net.sf.dftools.architecture.slam.component.ComponentPackage#getComNode_Speed()
 	 * @model default="1" required="true"
 	 * @generated
 	 */
-	int getSpeed();
+	float getSpeed();
 
 	/**
 	 * Sets the value of the '{@link net.sf.dftools.architecture.slam.component.ComNode#getSpeed <em>Speed</em>}' attribute.
@@ -78,5 +78,5 @@ public interface ComNode extends Component {
 	 * @see #getSpeed()
 	 * @generated
 	 */
-	void setSpeed(int value);
+	void setSpeed(float value);
 } // ComNode
